Add Facebook streaming controller and integrate with main application
